requirements-dev.txt
# Development and Testing Dependencies
# Install with: pip install -r requirements-dev.txt

# Testing
pytest>=7.4.0
pytest-django>=4.5.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
factory-boy>=3.3.0
faker>=19.0.0

# Code Quality
flake8>=6.0.0
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0
pre-commit>=3.3.0

# Security
bandit>=1.7.0
safety>=2.3.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0

# Development Tools
django-debug-toolbar>=4.2.0
django-extensions>=3.2.0
ipython>=8.14.0
